Simplify creation of webpage to which localhost resolves
Now the webpage is simply placed within the /var/www/html/ folder and it
is named index.html, and as such it takes precedence over the default
index.nginx-debian.html webpage with the domain 'localhost'. No need to
create a config file because the default one in
/etc/nginx/sites-available/default will handle listening at port 80 and
resolving for all domains that don't have their domain names set with
'server_name' in a config file.